refactor(aria/listbox): avoid circular references at runtime
Reworks the Aria listbox to avoid a circular dependency between `Listbox` and `Option`.
